rename top-level folder to unit-format
