+++ packages/nimble-components/src/table-column/number-text/index.ts
@@ -18,9 +18,9 @@ import { tableColumnNumberTextGroupHeaderTag } from './group-header-view';
 import { tableColumnNumberTextCellViewTag } from './cell-view';
 import type { ColumnInternalsOptions } from '../base/models/column-internals';
 import { NumberTextAlignment, NumberTextFormat } from './types';
-import type { UnitFormat } from '../../utilities/number-formatter/base/unit-format';
-import { DefaultUnitFormat } from '../../utilities/number-formatter/default-unit-format';
-import { DecimalUnitFormat } from '../../utilities/number-formatter/decimal-unit-format';
+import type { UnitFormat } from '../../utilities/unit-format/base/unit-format';
+import { DefaultUnitFormat } from '../../utilities/unit-format/default-unit-format';
+import { DecimalUnitFormat } from '../../utilities/unit-format/decimal-unit-format';
 import { TableColumnNumberTextValidator } from './models/table-column-number-text-validator';
 import { TextCellViewBaseAlignment } from '../text-base/cell-view/types';
 import { lang } from '../../theme-provider';
