+++ packages/nimble-components/src/table/specs/table-column-specs/table-column-formatted-text.md
@@ -233,9 +233,9 @@ Supporting a new unit requires defining a new `nimble-unit-<name>` element and a
 
 1.  Define a new unit class under `src/unit/<name>` that extends `Unit`. It must implement the method `getUnitScale()`, which should return the singleton instance of your new `UnitScale` class. You also must add all the standard code for registering the element with FAST, exporting its tag name, etc. See any of the existing unit elmements for an example.
 
-2.  Your `UnitScale` class goes under `src/utilities/number-formatter/unit-scale`. You must determine whether your unit is among [those supported by `Intl.NumberFormat`](https://tc39.es/ecma402/#table-sanctioned-single-unit-identifiers).
-    -   If it is, your new class should extend `UnitScale`, and you must implement the method `getSupportedScaledUnits()`. This should return an array of `IntlNumberFormatScaledUnitFormatter` instances. See `src/utilities/number-formatter/unit-scale/byte-unit-scale.ts` for an example.
-    -   If it is not, your class will be responsible for providing translated unit strings for a set of languages you choose to support. Your class should extend `ManuallyTranslatedUnitScale`. You must implement `getUnitTranslations()`, which will return a `ReadonlyMap` of [language tags](https://en.wikipedia.org/wiki/IETF_language_tag) to `UnitTranslation` objects. If you need different translations for the same language, based on region (e.g. "meter" for `en-US` vs. "metre" for `en-CA`), you may provide a locale string that includes the region (e.g. `en-CA`) as a key into your `ReadonlyMap`. You must also implement `getSupportedPrefixes()`, which returns an array of `UnitPrefix`. These prefixes define which scaled units (e.g. "kg", "mm", "GB") your unit is capable of converting to/displaying. For a standard set of SI prefixes, you may return `metricPrefixes`. See `src/utilities/number-formatter/unit-scale/volt-unit-scale.ts` for an example.
+2.  Your `UnitScale` class goes under `src/utilities/unit-format/unit-scale`. You must determine whether your unit is among [those supported by `Intl.NumberFormat`](https://tc39.es/ecma402/#table-sanctioned-single-unit-identifiers).
+    -   If it is, your new class should extend `UnitScale`, and you must implement the method `getSupportedScaledUnits()`. This should return an array of `IntlNumberFormatScaledUnitFormatter` instances. See `src/utilities/unit-format/unit-scale/byte-unit-scale.ts` for an example.
+    -   If it is not, your class will be responsible for providing translated unit strings for a set of languages you choose to support. Your class should extend `ManuallyTranslatedUnitScale`. You must implement `getUnitTranslations()`, which will return a `ReadonlyMap` of [language tags](https://en.wikipedia.org/wiki/IETF_language_tag) to `UnitTranslation` objects. If you need different translations for the same language, based on region (e.g. "meter" for `en-US` vs. "metre" for `en-CA`), you may provide a locale string that includes the region (e.g. `en-CA`) as a key into your `ReadonlyMap`. You must also implement `getSupportedPrefixes()`, which returns an array of `UnitPrefix`. These prefixes define which scaled units (e.g. "kg", "mm", "GB") your unit is capable of converting to/displaying. For a standard set of SI prefixes, you may return `metricPrefixes`. See `src/utilities/unit-format/unit-scale/volt-unit-scale.ts` for an example.
 
 For Angular and Blazor support, directives/wrappers will have to be created for your new unit element.
