+++ packages/nimble-components/src/unit/base/unit.ts
@@ -1,5 +1,5 @@
 import { FoundationElement } from '@microsoft/fast-foundation';
-import type { UnitScale } from '../../utilities/number-formatter/unit-scale/base/unit-scale';
+import type { UnitScale } from '../../utilities/unit-format/unit-scale/base/unit-scale';
 
 /**
  * An element representing a set of related units which could be used to represent the same value
