+++ packages/nimble-components/src/unit/byte/index.ts
@@ -1,9 +1,9 @@
 import { attr } from '@microsoft/fast-element';
 import { DesignSystem } from '@microsoft/fast-foundation';
 import { template } from '../base/template';
-import type { UnitScale } from '../../utilities/number-formatter/unit-scale/base/unit-scale';
-import { byte1024UnitScale } from '../../utilities/number-formatter/manually-translated-unit-scale/byte-1024-unit-scale';
-import { byteUnitScale } from '../../utilities/number-formatter/unit-scale/byte-unit-scale';
+import type { UnitScale } from '../../utilities/unit-format/unit-scale/base/unit-scale';
+import { byte1024UnitScale } from '../../utilities/unit-format/manually-translated-unit-scale/byte-1024-unit-scale';
+import { byteUnitScale } from '../../utilities/unit-format/unit-scale/byte-unit-scale';
 import { Unit } from '../base/unit';
 
 declare global {
