+++ packages/nimble-components/src/unit/byte/tests/unit-byte.spec.ts
@@ -1,8 +1,8 @@
 import { html } from '@microsoft/fast-element';
 import { type Fixture, fixture } from '../../../utilities/tests/fixture';
 import { UnitByte, unitByteTag } from '..';
-import { byte1024UnitScale } from '../../../utilities/number-formatter/manually-translated-unit-scale/byte-1024-unit-scale';
-import { byteUnitScale } from '../../../utilities/number-formatter/unit-scale/byte-unit-scale';
+import { byte1024UnitScale } from '../../../utilities/unit-format/manually-translated-unit-scale/byte-1024-unit-scale';
+import { byteUnitScale } from '../../../utilities/unit-format/unit-scale/byte-unit-scale';
 
 async function setup(binary: boolean): Promise<Fixture<UnitByte>> {
     return fixture<UnitByte>(html`
