+++ packages/nimble-components/src/utilities/unit-format/manually-translated-unit-scale/base/manually-translated-scaled-unit-format.ts
