+++ packages/nimble-components/src/utilities/unit-format/manually-translated-unit-scale/base/tests/manually-translated-unit-scale.spec.ts
